put the wait function inside of a general submission_utility.py script, which contains some of the other functions for submission, such as for changing the config
